feat: Moving queue file to it's directory and creating include file for it (#874)
* Putting queue in correct dir and creating include file for it

* Update data_structures/queue/include.h

missing one function, added
