+++ data_structures/queue/queue.c
@@ -1,35 +1,10 @@
 ////////////////////////////////////////////////////////////////////////////////
 // INCLUDES
-#include <stdio.h>
-#include <stdlib.h>
-
-////////////////////////////////////////////////////////////////////////////////
-// MACROS: CONSTANTS
-
-////////////////////////////////////////////////////////////////////////////////
-// DATA STRUCTURES
-/**
- * Defining the structure of the node which contains 'data' (type : integer), two pointers 'next' and 'pre' (type : struct node).
- */
-struct node                                          
-{
-    int data;
-    struct node *next;
-    struct node *pre;
-} * head, *tail, *tmp;
+#include "include.h";
 
 ////////////////////////////////////////////////////////////////////////////////
 // GLOBAL VARIABLES
 int count;
-
-////////////////////////////////////////////////////////////////////////////////
-// FORWARD DECLARATIONS
-void create();
-void enque(int x);
-int deque();
-int peek();
-int size();
-int isEmpty();
 
 ////////////////////////////////////////////////////////////////////////////////
 // MAIN ENTRY POINT
